added locks to prevent contention
Hardware level protection that is gained because of the protocol will not prevent the i2c registers from getting modified. These could be protected by the i2c driver but those locks are non-blocking, try-lock type in nature which will return error when the bus is busy. Thus it's a good idea to implement regular blocking locks on the driver functions.